refactor: Move launcher injects and stylesheets into modules/launcher folder
